requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
aiogram==2.2
python-dotenv==1.0.0